// src/FXChatBot/enums.ts
export enum BACKGROUND_COLOR {
    SECONDARY = "#2057A6",
}

export enum MESSAGE_SENDER {
    USER = "user",
    BOT = "bot"
}

export enum FONT_SIZE {
    MESSAGE = "14px",
    TITLE = "20px",
}

export enum INVESTMENT_CATEGORIES {
    MF = "Mutual Fund",
    STOCKS = "Stocks",
    GOLD = "Gold",
}

export enum FINANCIAL_ADVICE {
    INVESTMENT_PLAN = "Investment Plan",
    SPENDS_ADVICE = "Spends Advice",
    GOAL_PLANNING = "Goal Planning",
}

export enum MUTUAL_FUND {
    ABOUT = "About MF",
    TOP = "Top Performing Funds",
    FILTER = "Filter Funds",
}

export enum STOCKS {
    ABOUT = "About Stocks",
    TOP = "Top Performing Stocks",
    FILTER = "Filter Stocks",
}

export enum FILTER {
    RISK = "Risk",
    RETURN = "Return",
}

export enum RISK_TYPE {
    LOW = "Low",
    MEDIUM = "Medium",
    HIGH = "High",
}

export enum RETURN_RANGE {
    LOW = "Low",
    MEDIUM = "Medium",
    HIGH = "High",
}

export const WelcomeMessage = "Hello! I'll need some information to assist you better. Please fill out the form below.";
export const MF_INFO = "Mutual funds pool money from multiple investors to invest in a diversified portfolio of stocks, bonds, or other securities. They are managed by professional fund managers.";
export const STOCKS_INFO = "Stocks have the potential to provide high returns, especially over the long term. When you invest in stocks, you own a part of the company, and you can benefit from its growth and profits.";
export const GOLD_INFO = "Gold investment can be done in many forms like buying jewelry, coins, bars, gold exchange-traded funds, Gold funds, sovereign gold bond scheme, etc. Though there are times when markets see a fall in the prices of gold but usually it doesn't last for long and always makes a strong upturn.";